Rename syntect module
